refactor(record): consolidate set-related components into single files
- Combine create-set.usecase.ts and update-set.usecase.ts into set.usecase.ts
- Merge create-set.controller.ts and update-set.controller.ts into set.controller.ts
- Update import paths in API routes to use consolidated controller
- Adjust DI container configuration to reference new file structure
- Reorganize types.ts with improved grouping and comments

This change reduces file count while maintaining the same functionality,
improving code organization by grouping related components together.
