+++ src/infrastructure/di/types.ts
@@ -1,17 +1,24 @@
 export const TYPES = {
+  // repositories
   UserRepository: Symbol.for("UserRepository"),
   TeamRepository: Symbol.for("TeamRepository"),
   RecordRepository: Symbol.for("RecordRepository"),
 
+  // services
   AuthenticationService: Symbol.for("AuthenticationService"),
   AuthorizationService: Symbol.for("AuthorizationService"),
 
+  // usecases
   FindRecordUseCase: Symbol.for("FindRecordUseCase"),
   CreateRecordUseCase: Symbol.for("CreateRecordUseCase"),
+
   FindMatchesUseCase: Symbol.for("FindMatchesUseCase"),
+
   CreateSetUseCase: Symbol.for("CreateSetUseCase"),
-  CreateRallyUseCase: Symbol.for("CreateRallyUseCase"),
-  CreateSubstitutionUseCase: Symbol.for("CreateSubstitutionUseCase"),
   UpdateSetUseCase: Symbol.for("UpdateSetUseCase"),
+
+  CreateRallyUseCase: Symbol.for("CreateRallyUseCase"),
   UpdateRallyUseCase: Symbol.for("UpdateRallyUseCase"),
+
+  CreateSubstitutionUseCase: Symbol.for("CreateSubstitutionUseCase"),
 };
